// packages/hub-center/config/constants/stableSwap/types.ts
import type { SerializedWrappedToken } from '@plexswap/metalists'
import { Address } from 'viem'

export interface BasePool {
  lpSymbol: string
  lpAddress: Address
  token: SerializedWrappedToken
  quoteToken: SerializedWrappedToken
}

export interface StableSwapPool extends BasePool {
  stableSwapAddress: Address
  infoStableSwapAddress: Address
  stableLpFee: number
  stableLpFeeRateOfTotalFee: number
}